% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback-wrappers.R
\name{feedbackDanger}
\alias{feedbackDanger}
\title{feedbackDanger}
\usage{
feedbackDanger(inputId, condition, text = "Danger, turn back!",
  color = "#d9534f", icon = shiny::icon("exclamation-sign", lib =
  "glyphicon"), cancelOutput = FALSE)
}
\arguments{
\item{inputId}{the Shiny input's \code{inputId} argument}

\item{condition}{condition under which feeback is displayed}

\item{text}{default \code{"Danger, turn back"} text string to display below input}

\item{color}{defailt \code{"#d9534f"} the color of the feeback}

\item{icon}{default \code{icon("exclamation-sign", lib="glyphicon"))} a \code{shiny::icon} object}

\item{cancelOutput}{FALSE; whether or not to stop the reactive chain when
the \code{condition} argument evaluates to TRUE.  This calls 
\code{shiny::req(expr = !condition)} where \code{condition}
is the argument passed to \code{feedback}}
}
\description{
wrapper for \code{feedback()} function that displays a
danger message by default
}
\examples{

## Only run examples in interacive R sessions
if (interactive()) {
  ui <- fluidPage(
    useShinyFeedback(),
    
    numericInput(
      "exampleInput",
      "Show Feedback When < 0",
      value = -5
    )
  )
  
  server <- function(input, output) {
    observeEvent(input$exampleInput, {
      feedbackDanger(
        "exampleInput",
        condition = input$exampleInput < 0
      )
    })
  }
  
  shinyApp(ui, server)
}

}
\seealso{
\code{\link{feedback}}, \code{\link{feedbackWarning}}, \code{\link{feedbackSuccess}}

Other feedback.wrappers: \code{\link{feedbackSuccess}},
  \code{\link{feedbackWarning}}
}

