% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hideFeedback.R
\name{hideFeedback}
\alias{hideFeedback}
\title{hideFeedback}
\usage{
hideFeedback(inputId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{the Shiny input's \code{inputId} argument}

\item{session}{the \code{shiny} session}
}
\description{
hide feedback next to Shiny input
}
\examples{


}
\seealso{
\code{\link{showFeedbackWarning}}, \code{\link{showFeedbackDanger}}, \code{\link{showFeedbackSuccess}}
}
