% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadingButton.R
\name{loadingButton}
\alias{loadingButton}
\title{loadingButton}
\usage{
loadingButton(
  inputId,
  label,
  class = "btn btn-primary",
  style = "width: 150px;",
  loadingLabel = "Loading...",
  loadingSpinner = "spinner",
  loadingClass = NULL,
  loadingStyle = NULL
)
}
\arguments{
\item{inputId}{the input id}

\item{label}{the button text (label)}

\item{class}{the class(es) to apply to the button}

\item{style}{style for button (pre-loading); character string w/ CSS styling format: "color: black; background-color: red;"}

\item{loadingLabel}{text to show after button is clicked (e.g. during loading)}

\item{loadingSpinner}{the loading spinner icon.  Valid values are NULL, "spinner", "circle-notch",
"sync", and "cog"}

\item{loadingClass}{the loading button css class(es).}

\item{loadingStyle}{style for button (while loading); character string w/ CSS styling format: "color: black; background-color: red;"}
}
\description{
Button that becomes disabled until reset w/ `resetLoadingButton`
}
