% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeToast.R
\name{removeToast}
\alias{removeToast}
\title{Remove existing toast messages}
\usage{
removeToast(animate = FALSE, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{animate}{a logical indicating whether to remove the toast message(s) instantly or use its \code{hideMethod} with animations to remove (Default).}

\item{session}{the Shiny session. Defaults to \code{shiny::getDefaultReactiveDomain()}.}
}
\value{
`invisible()`
}
\description{
Remove existing toast messages
}
