% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadingButton.R
\name{resetLoadingButton}
\alias{resetLoadingButton}
\title{resetLoadingButton}
\usage{
resetLoadingButton(inputId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{the input id}

\item{session}{the shiny session}
}
\description{
Reset the `loadingButton` to its original style
}
