% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showFeedback.R
\name{showFeedback}
\alias{showFeedback}
\title{showFeedback}
\usage{
showFeedback(
  inputId,
  text = NULL,
  color = NULL,
  icon = NULL,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{inputId}{the Shiny input's \code{inputId} argument}

\item{text}{text string to display below input}

\item{color}{the color of the feedback}

\item{icon}{an html icon tag}

\item{session}{the \code{shiny} session}
}
\description{
show feedback next to Shiny input
}
\examples{

## Only run examples in interacive R sessions
if (interactive()) {
  ui <- fluidPage(
    useShinyFeedback(),
    
    numericInput(
      "exampleInput",
      "Show Feedback When < 0",
      value = -5
    )
  )
  
  server <- function(input, output) {
    observeEvent(input$exampleInput, {
      
      if (input$exampleInput < 0) {
      
        showFeedback(
          "exampleInput",
          text = "I am negative",
          color = "#d9534f",
          icon = shiny::icon("exclamation-sign", lib="glyphicon")
        )
      }
      
    })
  }
  
  shinyApp(ui, server)
}

}
\seealso{
\code{\link{showFeedbackWarning}}, \code{\link{showFeedbackDanger}}, \code{\link{showFeedbackSuccess}}
}
