% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showToast.R
\name{showToast}
\alias{showToast}
\title{show toast message}
\usage{
showToast(
  type,
  message,
  title = NULL,
  keepVisible = FALSE,
  .options = list(),
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{type}{length 1 character vector.  Valid values are "success", "error", "warning", and "info"}

\item{message}{the toast message}

\item{title}{the toast title.  Defaults to \code{NULL}}

\item{keepVisible}{a logical. If \code{TRUE}, the toast notification will remain visible until removed with \code{\link{hideToast}}. If \code{FALSE}, the default, the toast will automatically hide once the "showDuration" option has elapsed.}

\item{.options}{other options to pass to the \code{toastr} JavaScript library.  See 
\url{https://codeseven.github.io/toastr/demo.html} for a full demo of options.}

\item{session}{the Shiny session.  Defaults to \code{shiny::getDefaultReactiveDomain()}.}
}
\value{
`invisible()`
}
\description{
A wrapper around the `toastr` JavaScript library that uses our preferred default argument
values.
}
