% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snackbar-wrappers.R
\name{snackbarWarning}
\alias{snackbarWarning}
\title{snackbarWarning}
\usage{
snackbarWarning(id, message, includeRemoveButton = TRUE)
}
\arguments{
\item{id}{A length 1 character vector.  The unique id of the snackbar. This needs to match up with
the \code{id} argument specified in the serve side call to \code{showSnackbar()}.}

\item{message}{the text message to display in the snackbar}

\item{includeRemoveButton}{whether or not to show the remove button on the right side of the
snackbar}
}
\description{
This is a wrapper for \code{\link{snackbar}}.  It provides yellow warning styling for the snackbar.
}
\seealso{
\code{\link{showSnackbar}}

Other snackbar wrappers: \code{\link{snackbarDanger}},
  \code{\link{snackbarSuccess}}
}
\concept{snackbar wrappers}
