% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useShinyFeedback.R
\name{useShinyFeedback}
\alias{useShinyFeedback}
\title{useShinyFeedback}
\usage{
useShinyFeedback(feedback = TRUE, toastr = TRUE)
}
\arguments{
\item{feedback}{boolean: source in JS/CSS to use shinyFeedback functions (Default: \code{TRUE})}

\item{toastr}{boolean: source in JS/CSS to use showToast functions (Default: \code{TRUE})}
}
\description{
function to load js for using \code{shinyFeedback}
}
\section{Example}{

\preformatted{
  ui <- shinyUI(fluidPage(
    useShinyFeedback(
      feedback = TRUE,
      toastr = TRUE
    ),
    pageWithSidebar(
      headerPanel("Header"),
      sidebarPanel(
        ...
      ),
      mainPanel(
        ...
      )
    )
  ))
}
}

