% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valueBoxModule.R
\name{valueBoxModule}
\alias{valueBoxModule}
\title{valueBoxModule}
\usage{
valueBoxModule(input, output, session, value, subtitle = function() NULL)
}
\arguments{
\item{input}{the Shiny server input}

\item{output}{the Shiny server output}

\item{session}{the Shiny server session}

\item{value}{Either a reactive or an R object that can be coerced into a string.
The value to be displayed in the value box.}

\item{subtitle}{reactive to dynamically set the subtitle.  Set the "subtitle" argument
of \code{valueBoxModuleUI()} to "__server__" to display this subtitle.}
}
\description{
Server function for the `valueBoxModule`.  `valueBoxModule` is similar to 
`shinydashboard::valueBox()` but it moves the UI from the server to the ui (
i.e. the entire box is not rendered when the value in the value box updates;
only the actual value is rerendered).  By moving the box content to the UI 
the value box does not flash onto the screen when rendered.
}
\details{
`valueBoxModule` also allows for more custom styling of the box colors than
`shinydashboard::valueBox()`.
}
