% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valueBoxModule.R
\name{valueBoxModuleUI}
\alias{valueBoxModuleUI}
\title{valueBoxModuleUI}
\usage{
valueBoxModuleUI(
  id,
  subtitle,
  icon = NULL,
  backgroundColor = "#7cb5ec",
  textColor = "#FFF",
  width = 4,
  href = NULL,
  iconColor = grDevices::rgb(0, 0, 0, 0.15)
)
}
\arguments{
\item{id}{the Shiny module id}

\item{subtitle}{The subtitle to be displayed in the value box}

\item{icon}{An icon made by the `shiny::icon()`}

\item{backgroundColor}{A hex color code string}

\item{textColor}{A hex color code string}

\item{width}{A number between 1 and 12}

\item{href}{A url}

\item{iconColor}{A valid color string}
}
\description{
valueBoxModuleUI
}
