% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{feedback}
\alias{feedback}
\alias{feedbackWarning}
\alias{feedbackDanger}
\alias{feedbackSuccess}
\title{feedback (Defunct)}
\usage{
feedback(inputId, condition, text = NULL, color = NULL, icon = NULL)

feedbackWarning(inputId, condition, text = NULL, color = NULL, icon = NULL)

feedbackDanger(inputId, condition, text = NULL, color = NULL, icon = NULL)

feedbackSuccess(inputId, condition, text = NULL, color = NULL, icon = NULL)
}
\arguments{
\item{inputId}{the Shiny input's \code{inputId} argument}

\item{condition}{condition under which feedback is displayed}

\item{text}{text string to display below input}

\item{color}{the color of the feedback}

\item{icon}{a \code{shiny::icon} object}
}
\description{
These functions have been removed from `shinyFeedback`.  Use `\link{showFeedback}()` 
and `\link{hideFeedback}()` instead.
}
