% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback-wrappers.R
\name{feedbackDanger}
\alias{feedbackDanger}
\title{feedbackDanger}
\usage{
feedbackDanger(inputId, condition, text = "Danger, turn back!",
  color = "#d9534f", icon = shiny::icon("exclamation-sign", lib =
  "glyphicon"))
}
\arguments{
\item{inputId}{the Shiny input's \code{inputId} argument}

\item{condition}{condition under which feeback is displayed}

\item{text}{default \code{"Danger, turn back"} text string to display below input}

\item{color}{defailt \code{"#d9534f"} the color of the feeback}

\item{icon}{default \code{icon("exclamation-sign", lib="glyphicon"))} a \code{shiny::icon} object}
}
\description{
wrapper for \code{feedback()} function that displays a
danger message by default
}

