% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadingButton.R
\name{resetLoadingButton}
\alias{resetLoadingButton}
\title{reset_loading_button}
\usage{
resetLoadingButton(inputId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{the input id}

\item{session}{the shiny session}
}
\description{
Reset the `loading_button` to its original style
}
