% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snackbar.R
\name{showSnackbar}
\alias{showSnackbar}
\title{showSnackbar}
\usage{
showSnackbar(id, autoHideDuration = 3000)
}
\arguments{
\item{id}{A length 1 character vector.  A unique id for the snackbar.}

\item{autoHideDuration}{A length 1 numeric vector.  The abount of time in milliseconds to show the 
snackbar (e.g. `3000` is 3 seconds).  Set to NULL to keep snackbar open indefinitely.}
}
\description{
Server side function to show a snackbar.  This function should be called in the expression passed
to `shiny::observe()` or `shiny::observeEvent()`.
}
\examples{

## Only run examples in interacive R sessions
if (interactive()) {
  
  ui <- fluidPage(
    useShinyFeedback(),
    
    actionButton(
      "showSnackbarBtn",
      "Show Snackbar"
    ),
    
    snackbar(
      message = "Have a snack!"
    )
  )
  
  server <- function(input, output) {
    observeEvent(input$showSnackbarBtn, {
      showSnackbar("my_snackbar")
    })
  }
  
  shinyApp(ui, server)
}

}
