% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showToast.R
\name{showToast}
\alias{showToast}
\title{show toast message}
\usage{
showToast(
  type,
  message,
  title = NULL,
  .options = list(),
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{type}{length 1 character vector.  Valid values are "success", "error", "warning", and "info"}

\item{message}{the toast message}

\item{title}{the toast title.  Defaults to \code{NULL}}

\item{.options}{other options to pass to the `shinytoastr::toast_*()` function}

\item{session}{the Shiny session.  Defaults to \code{shiny::getDefaultReactiveDomain()}.}
}
\value{
`invisible()`
}
\description{
A wrapper around the `shinytoastr::toast_*()` functions that uses our preferred default argument
values.
}
