% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snackbar.R
\name{snackbar}
\alias{snackbar}
\title{snackbar}
\usage{
snackbar(id, message)
}
\arguments{
\item{id}{A length 1 character vector.  The unique id of the snackbar. This needs to match up with
the `id` argument specified in the serve side call to `showSnackbar()`.}

\item{message}{the text message to display in the snackbar}
}
\description{
UI side function to generate the snackbar html.
}
